rearrange index.js components and move refresh mutation to prevent reload on translate
